--- project/controllers/footer.html
@@ -1,1 +1,0 @@
-<footer></footer>
