--- project/controllers/header.html
@@ -1,4 +1,0 @@
-<header>
-    <img src="https://yu.ctu.edu.vn/images/upload/article/2020/03/0305-logo-ctu.png" class="logo"></i>
-    <h1>Quản lý sinh viên</h1>
-</header>
