--- project/controllers/sideBar.html
@@ -1,29 +1,0 @@
-<div id="sideBar">
-        <button onclick="hideNav()">===</button>
-        <ul>
-            <li>
-                <a href="../views/sinhvien.php">
-                    <img src="../assets/homeicon.png" class="icon"></img>
-                    <span>Quản Lý Sinh Viên</span>
-                </a>
-            </li>
-            <li>
-                <a href="../views/canbo.php">
-                    <img src="../assets/homeicon.png" class="icon"></img>
-                    <span>Thông Tin Cán Bộ</span>
-                </a>
-            </li>
-            <li>
-                <a href="../views/login.php">
-                    <img src="../assets/homeicon.png" class="icon"></img>
-                    <span>Đăng nhập</span>
-                </a>
-            </li>
-            <li>
-                <a href="../config/administrator.php">
-                    <img src="../assets/homeicon.png" class="icon"></img>
-                    <span>Đăng ký</span>
-                </a>
-            </li>
-        </ul>
-</div>
